Extract activity heartbeat component
